# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: RFQ Management System API for Cortoba Supplies
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /po:
    get:
      operationId: listPurchaseOrders
      tags: [po]
      summary: List all purchase orders
      parameters:
        - name: status
          in: query
          schema:
            type: string
        - name: search
          in: query
          schema:
            type: string
      responses:
        "200":
          description: List of purchase orders
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/PurchaseOrder"
    post:
      operationId: createPurchaseOrder
      tags: [po]
      summary: Create a new purchase order from a sheet PO number, with per-item supplier assignment
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PurchaseOrderInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PurchaseOrder"
  /po/{id}:
    get:
      operationId: getPurchaseOrder
      tags: [po]
      summary: Get purchase order by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Purchase order
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PurchaseOrder"
  /po/{id}/items:
    get:
      operationId: listPurchaseOrderItems
      tags: [po]
      summary: Get items for a purchase order
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Purchase order items
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/PurchaseOrderItem"
  /po/lookup/{poNo}:
    get:
      operationId: lookupPoFromSheet
      tags: [po]
      summary: Lookup items from a purchase order number (reads Google Sheets DATA sheet, column K)
      parameters:
        - name: poNo
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Purchase order items from sheet
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/SheetPoItem"
  /po/sheets/po-numbers:
    get:
      operationId: listSheetPoNumbers
      tags: [po]
      summary: List unique purchase order numbers found in the sheet (column K)
      parameters:
        - name: sheet
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Purchase order numbers
          content:
            application/json:
              schema:
                type: object
                required: [poNumbers]
                properties:
                  poNumbers:
                    type: array
                    items:
                      type: string

  # PRICING TOKEN (Supplier Portal)
...
components:
  schemas:
...
    PurchaseOrder:
      type: object
      required: [id, internalPoNo, sheetPoNo, status, createdAt]
      properties:
        id:
          type: integer
        internalPoNo:
          type: string
        sheetPoNo:
          type: string
        receiverName:
          type: ["string", "null"]
        receiverPhone:
          type: ["string", "null"]
        status:
          type: string
        employeeId:
          type: integer
        employeeName:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]
        itemCount:
          type: integer
        createdAt:
          type: string
        updatedAt:
          type: string
    PurchaseOrderInput:
      type: object
      required: [sheetPoNo, items]
      properties:
        sheetPoNo:
          type: string
        receiverName:
          type: string
        receiverPhone:
          type: string
        notes:
          type: string
        items:
          type: array
          minItems: 1
          items:
            $ref: "#/components/schemas/PurchaseOrderItemInput"
    PurchaseOrderItemInput:
      type: object
      required: [description]
      properties:
        itemId:
          type: string
        lineItem:
          type: string
        partNo:
          type: string
        description:
          type: string
        uom:
          type: string
        qty:
          type: ["number", "null"]
        referencePrice:
          type: ["number", "null"]
        supplierId:
          type: ["integer", "null"]
        taxIncluded:
          type: boolean
          default: false
    PurchaseOrderItem:
      type: object
      required: [id, poId, description]
      properties:
        id:
          type: integer
        poId:
          type: integer
        supplierId:
          type: ["integer", "null"]
        supplierName:
          type: ["string", "null"]
        itemId:
          type: ["string", "null"]
        lineItem:
          type: ["string", "null"]
        partNo:
          type: ["string", "null"]
        description:
          type: string
        uom:
          type: ["string", "null"]
        qty:
          type: ["number", "null"]
        referencePrice:
          type: ["number", "null"]
        taxIncluded:
          type: boolean
    SheetPoItem:
      type: object
      required: [description]
      properties:
        itemId:
          type: string
        lineItem:
          type: string
        partNo:
          type: string
        description:
          type: string
        uom:
          type: string
        qty:
          type: ["number", "null"]
        referencePrice:
          type: ["number", "null"]
        poNo:
          type: string

    # OFFERS / PRICING
...